+++ server/app/Http/Controllers/Api/TurnosController.php
@@ -6,64 +6,69 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 
+use App\Http\Controllers\Api\BaseApiController;
 use App\Library\ResponseEstructure;
 use Illuminate\Support\Facades\Validator;
 
 use App\Turno;
 
-class TurnosController extends Controller
+class TurnosController extends BaseApiController
 {
     public function index(Request $request)
     {
-        $response_estructure = new ResponseEstructure();
-        $response_estructure->set_response(false);
+        $response_estructure = $this->validateToken($request->input("jwt_token"));
 
-        $fechaDesde = $request->input("fechaDesde");
-        $fechaHasta = $request->input("fechaHasta");
-
-        $input= [
-            "fecha_desde"=>$fechaDesde,
-            "fecha_hasta"=>$fechaHasta,
-        ];
-
-        $rules = [
-            "fecha_desde"=>"required|required|date_format:d/m/Y",
-            "fecha_hasta"=>"required|required|date_format:d/m/Y",
-        ];
-
-        $validator = Validator::make($input, $rules);
-
-        if ($validator->fails()) {
+        if($response_estructure->get_login() == true)
+        {
             $response_estructure->set_response(false);
 
-            $errors = $validator->errors();
+            $fechaDesde = $request->input("fechaDesde");
+            $fechaHasta = $request->input("fechaHasta");
 
-            foreach ($errors->all() as $error) {
-                $response_estructure->add_message_error($error);
+            $input= [
+                "fecha_desde"=>$fechaDesde,
+                "fecha_hasta"=>$fechaHasta,
+            ];
+
+            $rules = [
+                "fecha_desde"=>"required|required|date_format:d/m/Y",
+                "fecha_hasta"=>"required|required|date_format:d/m/Y",
+            ];
+
+            $validator = Validator::make($input, $rules);
+
+            if ($validator->fails()) {
+                $response_estructure->set_response(false);
+
+                $errors = $validator->errors();
+
+                foreach ($errors->all() as $error) {
+                    $response_estructure->add_message_error($error);
+                }
             }
-        }
-        else
-        {
-            $fechaDesde = \DateTime::createFromFormat("d/m/Y",$fechaDesde);
-            $fechaHasta = \DateTime::createFromFormat("d/m/Y",$fechaHasta);
+            else
+            {
+                $fechaDesde = \DateTime::createFromFormat("d/m/Y",$fechaDesde);
+                $fechaHasta = \DateTime::createFromFormat("d/m/Y",$fechaHasta);
 
-            $turnos = DB::table("turnos")
-            ->select(
-                "turnos.*",
-                "clientes.nombre as clientes_nombre",
-                "clientes.apellido as clientes_apellido",
-                DB::raw("DATE_FORMAT(turnos.fecha_hora_desde,'%d/%m/%Y %H:%i') as fecha_desde_esp"),
-                DB::raw("DATE_FORMAT(turnos.fecha_hora_hasta,'%d/%m/%Y %H:%i') as fecha_hasta_esp"),
-                DB::raw("DATE_FORMAT(turnos.created_at,'%d/%m/%Y %H:%i') as fecha_creacion_esp")
-            )
-            ->leftJoin("clientes","clientes.id","=","turnos.cliente_id")
-            ->where(DB::raw("DATE_FORMAT(turnos.fecha_hora_desde,'%Y/%m/%d')"),">=",$fechaDesde->format("Y/m/d"))
-            ->where(DB::raw("DATE_FORMAT(turnos.fecha_hora_desde,'%Y/%m/%d')"),"<=",$fechaHasta->format("Y/m/d"))
-            ->orderBy("turnos.id","desc")
-            ->get();
+                $turnos = DB::table("turnos")
+                ->select(
+                    "turnos.*",
+                    "clientes.nombre as clientes_nombre",
+                    "clientes.apellido as clientes_apellido",
+                    DB::raw("DATE_FORMAT(turnos.fecha_hora_desde,'%d/%m/%Y %H:%i') as fecha_desde_esp"),
+                    DB::raw("DATE_FORMAT(turnos.fecha_hora_hasta,'%d/%m/%Y %H:%i') as fecha_hasta_esp"),
+                    DB::raw("DATE_FORMAT(turnos.created_at,'%d/%m/%Y %H:%i') as fecha_creacion_esp")
+                )
+                ->leftJoin("clientes","clientes.id","=","turnos.cliente_id")
+                ->where(DB::raw("DATE_FORMAT(turnos.fecha_hora_desde,'%Y/%m/%d')"),">=",$fechaDesde->format("Y/m/d"))
+                ->where(DB::raw("DATE_FORMAT(turnos.fecha_hora_desde,'%Y/%m/%d')"),"<=",$fechaHasta->format("Y/m/d"))
+                ->orderBy("turnos.id","desc")
+                ->get();
 
-            $response_estructure->set_data($turnos);
-            $response_estructure->set_response(true);
+                $response_estructure->set_data($turnos);
+                $response_estructure->set_response(true);
+            }
         }
 
         return response()->json($response_estructure->get_response_array());
@@ -71,48 +76,12 @@ class TurnosController extends Controller
 
     public function store(Request $request)
     {
-        $response_estructure = new ResponseEstructure();
-        $response_estructure->set_response(false);
+        $response_estructure = $this->validateToken($request->input("jwt_token"));
 
-        $cliente_id = strtolower(trim($request->input("cliente_id")));
-        $precio = strtolower(trim($request->input("precio")));
-        $fecha_hora_desde = strtolower(trim($request->input("fecha_hora_desde")));
-        $fecha_hora_hasta = strtolower(trim($request->input("fecha_hora_hasta")));
-
-        $fecha_hora_desde = \DateTime::createFromFormat("d/m/Y H:i",$fecha_hora_desde);
-        $fecha_hora_desde = $fecha_hora_desde->format("Y-m-d H:i:s");
-
-        $fecha_hora_hasta = \DateTime::createFromFormat("d/m/Y H:i",$fecha_hora_hasta);
-        $fecha_hora_hasta = $fecha_hora_hasta->format("Y-m-d H:i:s");
-
-        $response_estructure->set_response(true);
-
-        if($response_estructure->get_response() == true)
+        if($response_estructure->get_login() == true)
         {
-            $turno_row = new Turno();
-            $turno_row->cliente_id = $cliente_id ;
-            $turno_row->precio = $precio;
-            $turno_row->fecha_hora_desde = $fecha_hora_desde;
-            $turno_row->fecha_hora_hasta = $fecha_hora_hasta;
-            $turno_row->save();
+            $response_estructure->set_response(false);
 
-            $response_estructure->set_response(true);
-        }
-
-        return response()->json($response_estructure->get_response_array());
-    }
-
-    public function edit(Request $request)
-    {
-        $response_estructure = new ResponseEstructure();
-        $response_estructure->set_response(false);
-
-        $id = $request->input("id");
-
-        $turno_row = Turno::where("id",$id)->first();
-
-        if($turno_row)
-        {
             $cliente_id = strtolower(trim($request->input("cliente_id")));
             $precio = strtolower(trim($request->input("precio")));
             $fecha_hora_desde = strtolower(trim($request->input("fecha_hora_desde")));
@@ -124,13 +93,57 @@ class TurnosController extends Controller
             $fecha_hora_hasta = \DateTime::createFromFormat("d/m/Y H:i",$fecha_hora_hasta);
             $fecha_hora_hasta = $fecha_hora_hasta->format("Y-m-d H:i:s");
 
-            $turno_row->cliente_id = $cliente_id ;
-            $turno_row->precio = $precio;
-            $turno_row->fecha_hora_desde = $fecha_hora_desde;
-            $turno_row->fecha_hora_hasta = $fecha_hora_hasta;
-            $turno_row->save();
-
             $response_estructure->set_response(true);
+
+            if($response_estructure->get_response() == true)
+            {
+                $turno_row = new Turno();
+                $turno_row->cliente_id = $cliente_id ;
+                $turno_row->precio = $precio;
+                $turno_row->fecha_hora_desde = $fecha_hora_desde;
+                $turno_row->fecha_hora_hasta = $fecha_hora_hasta;
+                $turno_row->save();
+
+                $response_estructure->set_response(true);
+            }
+        }
+
+        return response()->json($response_estructure->get_response_array());
+    }
+
+    public function edit(Request $request)
+    {
+        $response_estructure = $this->validateToken($request->input("jwt_token"));
+
+        if($response_estructure->get_login() == true)
+        {
+            $response_estructure->set_response(false);
+
+            $id = $request->input("id");
+
+            $turno_row = Turno::where("id",$id)->first();
+
+            if($turno_row)
+            {
+                $cliente_id = strtolower(trim($request->input("cliente_id")));
+                $precio = strtolower(trim($request->input("precio")));
+                $fecha_hora_desde = strtolower(trim($request->input("fecha_hora_desde")));
+                $fecha_hora_hasta = strtolower(trim($request->input("fecha_hora_hasta")));
+
+                $fecha_hora_desde = \DateTime::createFromFormat("d/m/Y H:i",$fecha_hora_desde);
+                $fecha_hora_desde = $fecha_hora_desde->format("Y-m-d H:i:s");
+
+                $fecha_hora_hasta = \DateTime::createFromFormat("d/m/Y H:i",$fecha_hora_hasta);
+                $fecha_hora_hasta = $fecha_hora_hasta->format("Y-m-d H:i:s");
+
+                $turno_row->cliente_id = $cliente_id ;
+                $turno_row->precio = $precio;
+                $turno_row->fecha_hora_desde = $fecha_hora_desde;
+                $turno_row->fecha_hora_hasta = $fecha_hora_hasta;
+                $turno_row->save();
+
+                $response_estructure->set_response(true);
+            }
         }
 
         return response()->json($response_estructure->get_response_array());
@@ -138,33 +151,37 @@ class TurnosController extends Controller
 
     public function get(Request $request)
     {
-        $response_estructure = new ResponseEstructure();
-        $response_estructure->set_response(true);
+        $response_estructure = $this->validateToken($request->input("jwt_token"));
 
-        $id = $request->input("id");
-
-        $turno_row = $turnos = DB::table("turnos")
-        ->select(
-            "turnos.*",
-            "clientes.nombre as clientes_nombre",
-            "clientes.apellido as clientes_apellido",
-            DB::raw("DATE_FORMAT(turnos.fecha_hora_desde,'%d/%m/%Y %H:%i') as fecha_desde_esp"),
-            DB::raw("DATE_FORMAT(turnos.fecha_hora_hasta,'%d/%m/%Y %H:%i') as fecha_hasta_esp"),
-            DB::raw("DATE_FORMAT(turnos.created_at,'%d/%m/%Y %H:%i') as fecha_creacion_esp")
-        )
-        ->leftJoin("clientes","clientes.id","=","turnos.cliente_id")
-        ->where("turnos.id",$id)
-        ->get();
-
-        if($turno_row)
+        if($response_estructure->get_login() == true)
         {
-            $turno_row = $turno_row[0];
-            $response_estructure->set_data($turno_row);
-        }
-        else
-        {
-            $response_estructure->set_response(false);
-            $response_estructure->add_message_error("El turno no existe");
+            $response_estructure->set_response(true);
+
+            $id = $request->input("id");
+
+            $turno_row = $turnos = DB::table("turnos")
+            ->select(
+                "turnos.*",
+                "clientes.nombre as clientes_nombre",
+                "clientes.apellido as clientes_apellido",
+                DB::raw("DATE_FORMAT(turnos.fecha_hora_desde,'%d/%m/%Y %H:%i') as fecha_desde_esp"),
+                DB::raw("DATE_FORMAT(turnos.fecha_hora_hasta,'%d/%m/%Y %H:%i') as fecha_hasta_esp"),
+                DB::raw("DATE_FORMAT(turnos.created_at,'%d/%m/%Y %H:%i') as fecha_creacion_esp")
+            )
+            ->leftJoin("clientes","clientes.id","=","turnos.cliente_id")
+            ->where("turnos.id",$id)
+            ->get();
+
+            if($turno_row)
+            {
+                $turno_row = $turno_row[0];
+                $response_estructure->set_data($turno_row);
+            }
+            else
+            {
+                $response_estructure->set_response(false);
+                $response_estructure->add_message_error("El turno no existe");
+            }
         }
 
         return response()->json($response_estructure->get_response_array());
@@ -172,22 +189,26 @@ class TurnosController extends Controller
 
     public function delete(Request $request)
     {
-        $response_estructure = new ResponseEstructure();
-        $response_estructure->set_response(false);
+        $response_estructure = $this->validateToken($request->input("jwt_token"));
 
-        $id = $request->input("id");
-
-        $turno_row = Turno::where("id",$id)->first();
-
-        if($turno_row)
-        {
-            $turno_row->delete();
-            $response_estructure->set_response(true);
-        }
-        else
+        if($response_estructure->get_login() == true)
         {
             $response_estructure->set_response(false);
-            $response_estructure->add_message_error("El turno no existe");
+
+            $id = $request->input("id");
+
+            $turno_row = Turno::where("id",$id)->first();
+
+            if($turno_row)
+            {
+                $turno_row->delete();
+                $response_estructure->set_response(true);
+            }
+            else
+            {
+                $response_estructure->set_response(false);
+                $response_estructure->add_message_error("El turno no existe");
+            }
         }
 
         return response()->json($response_estructure->get_response_array());
